requirements.txt
streamlit
google-cloud-aiplatform
google-generativeai
Pillow
PyYAML
google-cloud-storage
google-auth
